requirements.txt
pandas==1.1.5
pyspark==3.1.1
svgwrite==1.4
wget==3.2
google-cloud-storage
numpy~=1.21.6
requests~=2.27.1